Render tasklist on /tasks route
