+++ tests/dummy/app/routes/workflow/org/thread/tasks.js
@@ -1,0 +1,14 @@
+import Route from '@ember/routing/route';
+
+
+export default class WorkflowOrgThreadTasksRoute extends Route {
+  async model() {
+    let { user, thread, participants } = this.modelFor('workflow.org.thread');
+
+    return {
+      user,
+      thread,
+      participants
+    }
+  }
+}
