+++ tests/dummy/app/templates/workflow/org/thread.hbs
@@ -1,12 +1,8 @@
 <Cardflow
   @model={{this.model}}
-  {{!-- @org={{@model.currentOrg}} --}}
-  {{!-- @actionSteps={{this.actionSteps}} --}}
-  {{!-- @milestones={{this.milestones}} --}}
-  {{!-- @currentMilestone={{this.currentMilestone}} --}}
-  {{!-- @updateProgress={{this.updateProgress}} --}}
-  {{!-- @projectTitle={{this.projectTitle}} --}}
-  {{!-- @catalog={{this.catalog}} --}}
-  {{!-- @expandAction={{action this.displayCatalog}} --}}
-  {{!-- @transitionToProduct={{this.transitionToProduct}} --}}
+  @activeTasks={{this.activeTasks}}
+  @userTasks={{this.userTasks}}
+  @assignedTasks={{this.assignedTasks}}
 />
+
+{{outlet}}
